+++ src/core/hle/kernel/memory.cpp
@@ -14,7 +14,6 @@
 #include "core/hle/kernel/vm_manager.h"
 #include "core/hle/result.h"
 #include "core/hle/shared_page.h"
-#include "core/mem_map.h"
 #include "core/memory.h"
 #include "core/memory_setup.h"
