+++ src/core/hle/kernel/memory.h
